If a scenario has a step table or text, use original logic
A step text or step table makes things look very weird.
They didn't look that great in the first place,
but adding the indentation makes it worse.

The solution for now is to only indent And like steps
when there are no tables or texts in any of the steps