Get rid of step indentation formatting
Now that we can format step keywords how we want with a CSS file,
the step indentation is less desireable, and having thought about it
again, I think it isn't a good idea anyway.

No where in Gherkin do And steps get indented separately,
and also the indentation can give the idea that they are lesser
steps, when really they are just there so make it read more easily.

We can add this back in and have it configurable, but my feeling
is that we can drop it altogher.