Implement parsers option for --display-name-from-dir flag
This works the same way as --url-from-tag, so it supports the same
setup now. This meant some refactoring since the parsers
are needed in two different places. I decided to get them
once when the tool is called. It also adjust the logic around
the url parser so that the parser is only called if needed,
which brings it in line with how the new dir parser works.