Add a test for dir to display name conversion
This is a unit test for the utils library function,
so there is still some logic not tested, specifically,
the logic to detirmine whether the function comes from
the command line or a parser is untested.

But we have already been content with not testing fully
end to end by actually calling the whole process with
the sphinx-gherkindoc command, so this sticks to that.
We can talk about adding e2e testing, but I don't think
it adds enough value for the amount of time we'll need to set
aside to implement it.

This commit also changes the scope of a fixture because
it was causing test order to matter, which eeek. It
doesn't appear to have any real affect on performance