Fix tests by pointing to specific pytest-bdd commit as dep
I made a change on the scenario-descriptions branch
to change whitespace removal in descriptions, which
is needed for some upcoming work, but that ends up
breaking tests in master because our expected rST
output is now different even though we didn't change anything.

This locks it into the commit in the branch that matches our expected
behavior.