Use latest commit in pytest-bdd scenario description branch
This also simplifies the tests. It is much easier to
just store the actual pytest-bdd expectout rst output
rather than copy the behave one, but then also dynamically
adjust the expected output in the code. I find it easier
to see what is actually changed in the output.

As this commit shows, there are whitespace changes in the
expected output because pytest-bdd is now not stripping description
lines.