Add test for raw descriptions
NOTE: I explicitly did not add the test for the behave parser
plus raw descriptions. Since behave does not preserve line breaks
or indentation, my feeling is that it isn't a very realistic case.
In other words, I think it is fine to try it, but if behave does
something weird such that it doesn't work, we don't claim it
works. I'm open to the idea of causing an error if the behave
parser and this new option are combined, but didn't do that.