Refactor get_all_included_scenarios logic
This function was very large and complicated.
I attempted to split it up into a few helper functions,
but even with this change, I'm not sure it is any better.
I made it in a separate commit so that reviewers can let me know
if they think the original large function is preferrable.